requirements.txt
Flask==3.1.1
openai==1.82.0
python-dotenv==1.1.0
gunicorn==23.0.0
SpeechRecognition==3.14.3